Start better delegation to services

+++ apps/api/src/app/socket/socket.gateway.ts
@@ -16,7 +16,6 @@ export class SocketGateway implements OnGatewayInit, OnGatewayConnection, OnGate
         private playerService: PlayerService,
         private matchService: MatchService,
         private socketService: SocketService,
-        private uuidService: UUIDService,
         private jwtService: JWTService
     ) { }
 
@@ -31,8 +30,14 @@ export class SocketGateway implements OnGatewayInit, OnGatewayConnection, OnGate
     handleDisconnect(@ConnectedSocket() client: Socket) {
         const token = client.handshake.query.token;
         const clientId = client.id;
-        const disconnectResult = this.socketService.disconnect(clientId, token);
-        const { player, payload, lobby } = disconnectResult;
+        const payload = this.jwtService.getPayload(token);
+        const player = this.playerService.getPlayer(payload.uuid);
+        if (!player) return;
+        const lobby = this.lobbyService.getLobby(player.lobby);
+        if (!lobby) return;
+        this.lobbyService.removePlayer(lobby, player);
+        this.playerService.deletePlayer(payload.uuid);
+        this.socketService.disconnect(clientId, token);
         this.matchService.removePlayer(player.lobby, payload.uuid);
         return lobby;
     }
@@ -41,7 +46,8 @@ export class SocketGateway implements OnGatewayInit, OnGatewayConnection, OnGate
     onEnterLobby(@MessageBody() lobbyQuery: { id: string }, @ConnectedSocket() client: Socket) {
         const token = client.handshake.query.token;
         const lobbyId = lobbyQuery.id;
-        return this.socketService.handleEnterLobby(lobbyId, token);
+        const lobby = this.lobbyService.getLobby(lobbyId);
+        return this.lobbyService.onEnterLobby(lobby, token);
     }
 
     @SubscribeMessage('ready')
@@ -53,16 +59,7 @@ export class SocketGateway implements OnGatewayInit, OnGatewayConnection, OnGate
         if (!player) return;
         const lobby = this.lobbyService.getLobby(player.lobby);
         if (!lobby) return;
-        player.ready = ready;
-        this.playerService.savePlayer(player);
-        const everyoneReady = this.isEveryoneReady(lobby);
-        if (everyoneReady) {
-            lobby.open = false;
-            this.lobbyService.saveLobby(lobby);
-            this.notifyStartGame(lobby);
-        } else {
-            this.socketService.notifyLobbyChanges(lobby);
-        }
+        this.lobbyService.onPlayerReady(lobby, player, ready);
     }
 
     @SubscribeMessage('throw dice')
@@ -74,21 +71,6 @@ export class SocketGateway implements OnGatewayInit, OnGatewayConnection, OnGate
         const match = this.matchService.getMatch(player.lobby);
         if (!match) return;
         await this.matchService.play(match, player);
-    }
-
-    private isEveryoneReady(lobby) {
-        for (const playerId of lobby.players) {
-            const p = this.playerService.getPlayer(playerId);
-            if (!p.ready) return false;
-        }
-        return true;
-    }
-
-    private notifyStartGame(_lobby) {
-        const lobby = cloneDeep(_lobby);
-        const players = lobby.players.map(id => this.playerService.getPlayer(id));
-        const match = cloneDeep(this.matchService.generateMatch(_lobby.board, ...players));
-        this.matchService.saveAndBroadcastMatch(match);
     }
 
 }
